Add (parent, title) contstrant for nodes
