+++ papermerge/core/signals.py
@@ -20,7 +20,6 @@ from django.conf import settings
 from papermerge.core.models import (
     Document,
     DocumentVersion,
-    Folder,
     User,
 )
 from papermerge.core.storage import get_storage_instance
@@ -82,30 +81,6 @@ def user_init(sender, instance, created, **kwargs):
     if created:
         if settings.PAPERMERGE_CREATE_SPECIAL_FOLDERS:
             instance.create_special_folders()
-
-
-@receiver([post_delete, post_save], sender=Document)
-@receiver([post_delete, post_save], sender=Folder)
-def if_inbox_then_refresh(sender, instance, **kwargs):
-    """
-    Inform inbox_refresh channel group that user's inbox was updated
-    """
-    # Folder or Document instance was deleted/moved from//to user's Inbox folder
-    try:
-        instance.refresh_from_db()
-    except (Document.DoesNotExist, Folder.DoesNotExist):
-        logger.warning('Too late - Document/Folder was already deleted')
-        return
-
-    try:
-        if instance.parent and instance.parent.title == Folder.INBOX_TITLE:
-            channel_layer = get_channel_layer()
-            async_to_sync(channel_layer.group_send)(
-                "inbox_refresh",
-                {"type": "inbox.refresh", "user_id": str(instance.user.pk)}
-            )
-    except Exception as ex:
-        logger.warning(ex, exc_info=True)
 
 
 def get_channel_data(task_name, type):
