+++ papermerge/notifications/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class NotificationsConfig(AppConfig):
     name = 'papermerge.notifications'
     label = 'notifications'
+
+    def ready(self):
+        from papermerge.core import signals  # noqa
